not killing cesuper.sh when running update from within cesuper.sh

+++ ce_main_app/shellscripts/ce_stop.sh
@@ -1,17 +1,8 @@
 #!/bin/sh
 
 ce_is_running() {
-    # find out whether we're on Yocto or Raspbian
-    issue=$( cat /etc/issue | grep -o "Yocto" | wc -l )
-
-    # If at least once the Yocto was found, it's Yocto
-    if [ "$issue" -gt "0" ]; then   # on Yocto
-        # Count how many instances are running
-        cnt=$( ps | grep cosmosex | grep -v grep | wc -l )
-    else                            # on Raspbian
-        # Count how many instances are running
-        cnt=$( ps -A | grep cosmosex | wc -l )
-    fi
+    # Count how many instances are running
+    cnt=$( ps -A | grep cosmosex | wc -l )
 
     # CosmosEx app is not running?
     if [ "$cnt" -eq "0" ]; then
@@ -22,9 +13,13 @@ ce_is_running() {
     return 1
 }
 
-# first kill the cesuper.sh script so it won't restart cosmosex app
-echo "Terminating cesuper.sh"
-killall -9 cesuper.sh > /dev/null 2>&1
+# kill the cesuper.sh script so it won't restart cosmosex app
+# don't kill cesuper.sh if requested, e.g. when called ce_main_app starts update, which runs ce_update.sh, which runs within cesuper
+# (then it only stops possibly running cosmosex app, but not the other scripts which might have cesuper.sh as parent in shell)
+if [ "$2" != "dontkillcesuper" ]; then
+    echo "Terminating cesuper.sh"
+    killall -9 cesuper.sh > /dev/null 2>&1
+fi
 
 # check if ce is running
 ce_is_running
@@ -36,26 +31,20 @@ if [ "$?" -eq "0" ]; then
 fi
 
 #------------------------
-# check for distro, run systemctl only on stretch
-distro=$( /ce/whichdistro.sh )
+# if should terminate (also) using systemctl, do it
+if [ "$1" != "nosystemctl" ]; then
+    # try to find out if cosmosex.service is installed
+    sysctl=$( systemctl status cosmosex )
+    sc_notfound=$( echo $sysctl | grep 'Loaded: not-found' | wc -l )
 
-if [ "$distro" = "stretch" ]; then
-    # if should terminate (also) using systemctl, do it
-    if [ "$1" != "nosystemctl" ]; then
-        # try to find out if cosmosex.service is installed
-        sysctl=$( systemctl status cosmosex )
-        sc_notfound=$( echo $sysctl | grep 'Loaded: not-found' | wc -l )
-
-        # if the cosmosex.service is installed (the not-found string was not found)
-        if [ "$sc_notfound" -eq "0" ]; then
-            # try to stop it using systemctl
-            systemctl stop cosmosex
-        fi
+    # if the cosmosex.service is installed (the not-found string was not found)
+    if [ "$sc_notfound" -eq "0" ]; then
+        # try to stop it using systemctl
+        systemctl stop cosmosex
     fi
 fi
 
 #------------------------
-
 # send SIGINT to allow terminate the app nicely
 echo "Terminating cosmosex - SIGINT"
 killall -2 cosmosex > /dev/null 2> /dev/null
